+++ src/main/java/insane96mcp/progressivebosses/module/dragon/DragonFeature.java
@@ -44,11 +44,12 @@ public class DragonFeature extends Feature {
     @Config
     @Label(name = "Enable Fixes", description = """
             Enable some fixes for the Ender Dragon:
-             - Dragon Head and Neck have been repositioned correctly
-             - Dragon will now play the growl sound only 4 times/second when respawning and when at the center breathing instead of 20/second (so your ears shouldn't blow up anymore)
+             - Dragon Head and Neck have been repositioned/resized correctly.
+             - Dragon will now play the growl sound only 4 times/second when respawning and when sitting instead of 20/second (so your ears shouldn't blow up anymore)
              - When the crystals that respawn the dragon in the center are destroyed, the fire is extinguished
              - Ender Dragon can now rise and fall faster (somewhere around 1.14 the multiplier for the y speed was reduced to 0.01 instead of 0.1 https://bugs.mojang.com/browse/MC-272431)
-             - Dragon is set exactly at the center of the well""")
+             - Dragon is moved exactly at the center of the well when landed
+             - Fixes entities accumulating knockback when hit by the dragon and then launching like a rocket""")
     public static Boolean enableFixes = true;
 
     @Config