+++ src/main/java/insane96mcp/progressivebosses/module/dragon/phase/CrystalRespawnPhase.java
@@ -91,7 +91,6 @@ public class CrystalRespawnPhase extends AbstractDragonPhaseInstance {
 		this.targetLocation = null;
 		this.spikesToRespawn.clear();
 		this.tick = 0;
-		this.dragon.sittingDamageReceived = this.dragon.getMaxHealth() * 0.10f;
 	}
 
 	/**