+++ changelog.md
@@ -1,6 +1,5 @@
 # Changelog
 
-* Dragon should try respawning crystal at the end of a phase
 * Dragon breath attack while flying
 * Minions should shoot like blazes
 * Speed up sitting phase
@@ -21,6 +20,7 @@
     * Fixed Ender Dragon growl sound playing 20 times (making it really loud) when about to respawn and when at the center breathing (it's now 4 times)
     * Fixed Ender Dragon growl sound being played in a too small radius making it inaudible from ground level
     * Fixed head and neck placements and head size
+    * Fixes entities accumulating knockback when hit by the dragon and then launching like a rocket
 * Rebalancements
   * Wither Charge Attack now ignores armor
   * Ultimate wither charge attack damage slightly increased (12 -> 13)